+++ ironfish/src/rpc/routes/wallet/importAccount.ts
@@ -3,8 +3,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import * as yup from 'yup'
 import { DecodeInvalidName, MultisigSecretNotFound } from '../../../wallet'
-import { decodeAccount } from '../../../wallet/account/encoder/account'
-import { BASE64_JSON_MULTISIG_ENCRYPTED_ACCOUNT_PREFIX } from '../../../wallet/account/encoder/base64json'
+import { decodeAccount } from '../../../wallet/exporter/account'
+import { BASE64_JSON_MULTISIG_ENCRYPTED_ACCOUNT_PREFIX } from '../../../wallet/exporter/base64json'
 import { DuplicateAccountNameError } from '../../../wallet/errors'
 import { RPC_ERROR_CODES, RpcValidationError } from '../../adapters'
 import { ApiNamespace } from '../namespaces'
