+++ ironfish/src/wallet/exporter/account.ts
@@ -1,8 +1,8 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { Assert } from '../../../assert'
-import { AccountImport } from '../../walletdb/accountValue'
+import { Assert } from '../../assert'
+import { AccountImport } from '../walletdb/accountValue'
 import { Base64JsonEncoder } from './base64json'
 import { Bech32Encoder } from './bech32'
 import { Bech32JsonEncoder } from './bech32json'
