+++ ironfish/src/wallet/exporter/bech32.test.ts
