+++ ironfish/src/wallet/exporter/bech32json.ts
@@ -1,8 +1,8 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { Bech32m } from '../../../utils'
-import { AccountImport } from '../../walletdb/accountValue'
+import { Bech32m } from '../../utils'
+import { AccountImport } from '../walletdb/accountValue'
 import { AccountDecodingOptions, AccountEncoder, DecodeFailed } from './encoder'
 import { JsonEncoder } from './json'
 export class Bech32JsonEncoder implements AccountEncoder {
