+++ ironfish/src/wallet/exporter/encoder.ts
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { multisig } from '@ironfish/rust-nodejs'
-import { LanguageKey } from '../../../utils'
-import { AccountImport } from '../../walletdb/accountValue'
+import { LanguageKey } from '../../utils'
+import { AccountImport } from '../walletdb/accountValue'
 
 export class DecodeInvalid extends Error {}
