+++ ironfish/src/wallet/exporter/spendingKey.ts
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { generateKeyFromPrivateKey, Key } from '@ironfish/rust-nodejs'
-import { AccountImport } from '../../walletdb/accountValue'
-import { ACCOUNT_SCHEMA_VERSION } from '../account'
+import { AccountImport } from '../walletdb/accountValue'
+import { ACCOUNT_SCHEMA_VERSION } from '../account/account'
 import {
   AccountDecodingOptions,
   AccountEncoder,
