+++ ironfish/src/wallet/index.ts
@@ -3,10 +3,10 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 export * from './account/account'
 export * from './wallet'
-export * from './account/encoder/encoder'
-export * from './account/encoder/account'
+export * from './exporter/encoder'
+export * from './exporter/account'
 export { AccountValue } from './walletdb/accountValue'
-export { Base64JsonEncoder } from './account/encoder/base64json'
-export { JsonEncoder } from './account/encoder/json'
+export { Base64JsonEncoder } from './exporter/base64json'
+export { JsonEncoder } from './exporter/json'
 export * from './validator'
 export * from './walletdb/walletdb'
