+++ ironfish/src/wallet/walletdb/accountValue.ts
@@ -5,7 +5,7 @@ import { PUBLIC_ADDRESS_LENGTH } from '@ironfish/rust-nodejs'
 import bufio from 'bufio'
 import { IDatabaseEncoding } from '../../storage'
 import { ACCOUNT_KEY_LENGTH } from '../account/account'
-import { MultisigKeysEncoding } from '../account/encoder/multisigKeys'
+import { MultisigKeysEncoding } from '../exporter/multisigKeys'
 import { MultisigKeys, MultisigKeysImport } from '../interfaces/multisigKeys'
 import { HeadValue, NullableHeadValueEncoding } from './headValue'
